Move RenderMarkdown dir
