+++ apps/trpfrog.net/src/markdown/code-block-components/definitions/result-box/ResultBox.tsx
